--- a/.dockerignore
+++ b/.dockerignore
@@ -1,21 +1,73 @@
+# Dependencies
 node_modules
+**/node_modules
+
+# Build outputs
 .next
+**/.next
 dist
-.git
-.gitignore
-README.md
+**/dist
+.turbo
+**/.turbo
+out
+**/out
+
+# Test artifacts
+coverage
+**/coverage
+.nyc_output
+playwright-report
+test-results
+e2e/downloads
+e2e/screenshots
+
+# Environment files
 .env
 .env.*
 !.env.example
+
+# Version control
+.git
+.gitignore
+.github
+
+# IDE
 .vscode
 .idea
-coverage
-.nyc_output
+*.swp
+*.swo
+*~
+
+# OS
+.DS_Store
+Thumbs.db
+
+# Logs
+*.log
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
-.DS_Store
-*.log
+pnpm-debug.log*
+lerna-debug.log*
+
+# Documentation
+README.md
+docs
+*.md
+!packages/*/README.md
+
+# Temporary files
 tmp
 temp
-.turbo
+*.tmp
+
+# Development files
+.husky
+.eslintcache
+*.tsbuildinfo
+tsconfig.tsbuildinfo
+
+# Docker files (avoid recursive builds)
+Dockerfile*
+docker-compose*.yml
+.dockerignore